chore: move tests to tests.rs
